+++ CHANGELOG.md
@@ -1,9 +1,8 @@
 # Unreleased
 
-- **[NEW]** `SeaSection` holds SEA Page layout
-
 [...]
 
+- **[NEW]** `SeaSection` holds SEA Page layout
 - **[NEW]** Add `Marquee`
 - **[NEW]** Add `TripCardSample`